[core-coq] Fix implicit arguments for constructors
This change comes in two parts:

1. In the preamble, we include the magic lines

        Set Implicit Arguments.
        Unset Strict Implicit.
        Unset Printing Implicit Defensive.

2. For all nullary constructors `C : T a1 … aN` where `N > 0`, we
   make the parameters implicit:

        Arguments C {_} … {_}.

   where there are `N` `{_}`s.

Without these, when we declare inductive data types, their parameters
are *explicit* arguments to their constructors.  So if we have

    Inductive Maybe (A : Type) : Type
      := Nothing : Maybe A
      |  Just    : A -> Maybe A.

    Inductive Option (A : Type) : Type
      := None : Option A
      |  Some : A -> Option A.

then we must write

    Definition to_option {A} (m : Maybe A) : Option A :=
      match m with
        | Nothing _   => None _
        | Just    _ x => Some _ x
      end.

where each of the `_`s corresponds to `A`.  This, of course, breaks the
translation, and is quite silly to boot.  The first change (`Set
Implicit Arguments`) makes the arguments to non-nullary constructors
automatically implicit, so we can write

    Definition to_option {A} (m : Maybe A) : Option A :=
      match m with
        | Nothing _   => None _
        | Just      x => Some   x
      end.

However, there's no way to make this behavior automatic for nullary
constructors, so we need the second fix; the explicit lines

    Arguments Nothing {_}.
    Arguments None    {_}.

are enough to enable us to write the desired

    Definition to_option {A} (m : Maybe A) : Option A :=
      match m with
        | Nothing   => None
        | Just    x => Some x
      end.