Implement `skip method` (fixes #17)
The normal `skip` is a bit to weak, because a skipped method needs to be
skipped both in the class definition and the instance definition.

Format:

    skip method Monad fail

The `Monad` is unqualified; this seemed to be the most convenient
choice, given that this edit needs to be used both where a class is
declared (and hence the name is unqualified) and where a class is
referenced (where it can be qualified).