Support for deriving instances
This seems to work rather well. At least I had to `skip` many more
instances in `base-ghc`…

The code still needs to move to its own module. And it is quite a
hack! (but a 10-line patch extending the GHC-API would make that go
away).

This patch also includes:
Do not axiomatize failing instances instead comment them out. The Axioms
themself can fail to typecheck, and they cannot be skipped.