Disable Implicit Arguments, and set `Arguments` for all constructors
After reading the docs I begin to believe that we do not want that. All
arguments that should be implicit are marked as such by us (`{`…`}`), so
clearly we don’t want Coq to arbitrarily make other arguments implicit.
Should fix #20.

It requires settings `Arguments` for all data constructors. We nicely
have the infrastructure.