Calculate required `Require` sentences correctly
Previously, if you have `Foo.Bar.baz` in the code, it would create

    Require Foo
    Require Foo.Bar

which is wrong. The reason was that generic programming was used to
collect all qualified ids, to then take the module name, but `Foo.Bar`
itself is a qualified id, which yields the unwanted `Foo`.

By using `getFreeVars` instead of the generic `listify` fixes this.

This patch also simplifies some code in Lexing which I first thought was
the reason for this, although it turned out it was not.

A few `skip module GHC` edits are now no longer necessary.