Ensure that Qualid names are indeed qualified
and that the `AccessIdent` does not contain a module prefix. I am using
pattern synonyms that enforce this invariant to catch cases where it
is violated (and `infixToCoq` was one such case), but plan to revert
back to normal constructors once the refactoring is completeld (partly
because ghc-8.0 does not yet support `COMPLETE` pragmas).