Support operator names in patterns, by z-encoding them.
This allows the translation of `Data.Function.on`.

It raises the question, however, when we start supporting notation again
how we will know whether an operator is top-level (presumably with
notation) or local (no notation avilable). But later down the line, I
expect top-level names to be always qualified in the AST, so that should
work.