Checkpoint: switching Indent to Qualid in the translation
This compiles (yay) but is, in this form, horribly broken, because I did
not touch the AST yet. Since we use the Gallina AST not just in the
output, but also as internal data structure, there are many bogus calls
to `Bare` and `qualidBase` as things move in and out the AST.

We need to retain the required information (qualified names) in the AST
as well. So I will tackle that next.

The ultimate goal is:
 * Every top level entitiy is always and consistenly addressed using
   its qualified name.
 * Upon parsing, unqualified names need to be qualified.
   (Mostly for edits and built-in information)
 * If that is too tedious, the parsing needs to know the current module
   and qualify things as needed.
 * We keep the qualified name until pretty-printing. Then either
    * we de-qualify names with the current module names just before
      pretty-printing (e.g. using a generic traversal), or
    * The pretty-printer knows the current module and de-qualifies
      on-the-go.

What about locally bound variables (e.g. parameters)? Under this scheme
it should work both if they are qualified or bare. I will see which
works better the code.

Similarly for dependent names like type class methods: Qualified or
bare? My hunch is: Interally qualified is better.