Do not treat nullary constructors like variables
Previously, the constructor `ArgsPat` was only used for constructors
applied to arguments, while both nullary constructors and variables were
`QualidPat`. This represents syntax well, but not meaning, and
for code that we get from GHC we know what is a constructor, and we want
to preserve this information in the AST. Hence I now put nullary
constructors into `ArgsPat` (with an empty argument list).

This fixes #60.

This is not a complete refactoring: we sometimes use `QualidPat`
internally for hardcoded nullary constructors, and in particular the Coq
parser for edits will have a hard time distinguishing `null` from `nul`.
So I left the (only partially working, see #60) code in place that
checks `QualidPat` for constructors. But this is clearly a technical dept...