Partial work on proving ordering.
The current approach is very difficult because it requires a lot of lemmata. It would be nice if I have a property showing that [ordered] respects equivalence relation of elements as well as that of trees, but I haven't figured out how to do that. The other possible direction that may simplify the work is to define an inductive type that respects [ordered] property so we can do proof by induction on ordering relations.