More work on IntSet, but now hit a serious road block
I realized that when we combine two trees, we need to know the highest
bit where they differ. This would be fine with Int64, or Word64 or even
N, but not with Z: -1 has infinitely many bits set to 1, while 0 has
infinitely many bits set to 0. So we cannot do that. That is
unfortunate.

One way out would be to change the Bits instance for Int to not have
two-complement’s representation, and store the sign in the lowest bit.
But that diverges too much from both `containers` and what’s provided
in the Coq library (which has a pretty decent standard library for bit
manipulation on Z and N).

I did not see fixed-width integer library in the standard libraries.

One way out would be to set

    Definition elt := {z : Z | is64BitWide z}

instead of

    Definition elt := Z.

and thus declare our interface to only be compatible with 64-bit
integers.