A stab at converting Data.Sequence.
We're are a long way from converting this file automatically. The biggest issue is that
our treatment of polymorphic functions does not permit polymorphic recursion by default.

That said, there are still a number of blockers:
   - I've converted almost all of the GHC.Num.Ints to nats to make termination reasoning
     easier. This also means converting the operations too: in the future we can add
     Num and Ord instances for nat.
   - There are several places where Coq's type inference is not up-to-par. I've had to
     add extra type annotations and rewrite the generated output just to get things
     to type check. A particularly difficult one was `applicativeTree` which causes Coq
     to StackOverflow (in its generated version).
   - The manual version is not done. I still can't get some of it to typecheck.
   - Mutual recursion.

I did unwind all of the type class ordering dependencies. Those are recorded in the edit
file.

If we ever finish this, we should compare to Matthieu Sozeau's version of fingertrees in Coq.