Refactor story around obligations and use of Program mode
obligations are now specified in a separte `obligations` edit. In
obligation mode (which is enabled either by `termination` or
`obligations`), if-then-else is desugared to a case statement, to get
useful termination principles.

See base-test/SplitAt for a good example of local non-structural
recursion with a proof.

Small wart: Since we use Program mode for instances, we now print `Admit
All Obligations` after them. Since there are no obligations, it does not
matter for now (but I can revisit that point later).