Switch EqExact back to a subclass of EqLaws
However, now there's Ltac to make your life easier.  `EqLaws` follows from
`EqExact` plus the correspondence between `==` and `/=`, so we have

    Lemma Build_EqLaws_reflect (t : Type) `{Eq_ t} :
      (forall x y : t, reflect (x = y) (x == y)) ->
      (forall x y, x == y = ~~ (x /= y))         ->
      EqLaws t.

There's then the Ltac command `EqLaws_from_reflect Eq_eq_t` which, given a proof
of the first reflection property, will attempt to use `Build_EqLaws_reflect` to
produce an `EqLaws` instance; it will automatically prove the `==`-`/=`
correspondence if `/=` is using the default definition.  The usage pattern is as
follows:

    Theorem Eq_eq_Int (x y : Int) : reflect (x = y) (x == y).
    Proof.
      (* ... *)
    Qed.

    Instance EqLaws_Int : EqLaws Int.
    Proof. EqLaws_from_reflect Eq_eq_Int. Qed.

    Instance EqExact_Int : EqExact Int.
    Proof. constructor; apply Eq_eq_Int. Qed.

If `EqLaws_from_reflect` can't fill in the `==`-`/=` correspondence, it will
leave that as a goal.