Verify big chunks of `Data.Map`
we have: `singleotn`, `balanceL`, `balance_R`, `insertMin`, `insertMax`,
`link`, `insert`, `insertR`, `maxViewSure`, `minViewSure`, `glue`,
`delete`, `split`, `union`, `link2`, `splitMember, `intersection`,
`difference`.

That should be enough for us to claim that we have three modules.

The `f_solver` tactic does not work quite as well as in with `Set`, and
I left `TODO` in ugly manual proofs that should be handled by the
tactics. It is probably mostly teaching `f_solver` about the
option-valued equivalents of `||` and `&&`.

Unfortunately it is quite slow. I think this mostly the unrolling lemmas
for Program Fixpoint-defined things :-(