Remove __uniqe from ConversionState (and use CounterT instead)
this is working towards clearning out the ConversionMonad type class, to
clearly distinguish what is state, what is envionment, and eventually,
which part of the state needs to be written to the interface files.

As a first step, the counter (which should exist only be per-definition
anyways), is pulled out.

En passant I removed some complexity that seems to stem from an attempt
at handling multiple modules when were working pre-renamer.