Put all non-mutable state in a reader monad
and distinguish between
 * GlobalMonad (has edits)
 * ConversionMonad (dito + in the scope of one module)
 * LocalConvMonad  (dito + current definition and uniq counter)

`mtl` itself does not support an extensible `MonadReader`, so I shopped
aorund a bit, and settled for a solution based on classy lenses, as
provided by the `lens` package. Works nice so far.

Biggest downside now is that the `GlobalMonad` now has a spurious `r`
type parameter, which is unfortunate, as there is a functional
dependency that determines `r` from `m`...