Refactoring of how free variables are calculated
going from a monad transformer to a simple monoid with combinators. It
seems that we are not useing the full abstract power of a monad
transformer, so let’s keep it simple.

Module `HsToCoq.Util.FVs` contains type classes and combinators
to work with
 * types that have free variables (`HasFV`) as well as
 * types that have free variables, but also bind something, i.e. export
   a name (`HasBV`).
It has combinatos to combine the two, such as `scopesOver` or
`telescope` or `scopesMutually`, and to convert between them
(`bindsNothing`, `forgetBinders`).

The codes uses `TypeError` to signal “dangerous“ instances, such as
lists of binders – do they scope over each other (i.e. a telescope), or
do they scope not (as in patterns), to force me to think whether I want
to use `foldMap` or `foldScopes` or `scopesMutually`. This proved
useful. Maybe `foldTelescope` would be a convenient default, though.

`HsToCoq.Coq.FreeVars` contains, as before, the instances for the AST.

The output diff has identical code, but shows two behaviuor changes in
the `Exported Variables` section:
 * Record field name are collected (no big deal)
 * The old code must have had a bug, because previously, `id`
   was lost as a free variable in `ghc/lib/CoreSyn.v`, probably
   because it was a parameter to a `Inductive`.
   So hey, the refactoring was not just an exercise in futility, and it
   adds confidence that a simpler combinator-based interface works
   better.

Of course this can be further improved; maybe some more common idiom
should be put in combinators (`_ <> bindsNothing _` comes to mind), and
I am not sure if we really need both `scopesOver` and `telescope`.