Instance translation refactoring
This moves the heurisitcs as to whether a method call should be
translated to a method call, or to a direct call to the instance’s method,
much earlier in the pipeline, i.e. making it behave like “rename” edits.

This gets rid of the rather complicate substition code.

It allows `in Foo.Class__Ty_metho` edits to work properly.

Because local edit override these global edits, this allows edits like
```
in Control.Monad.Trans.Writer.Lazy.Applicative__WriterT_op_zlztzg__ rename value GHC.Base.liftA2 = GHC.Base.liftA2
```
to override the heuristics. This should allow us to generate `Eq []`,
for example.

And, as a side-effect, this should work nicely with mutual recursion;
i.e. be mostly orthotogonal to it.