Mark test `MultrecInst` as failing
previously, we would just have a comment about unsupported mutual
recursion here, so this test would fail unwarrantedly. Now we produce
output, but Coq fails. This might just work once the `multrec` branch
lands (at least if the call to `addRecursion` for the top-level
definitions is moved to `Module.hs`)