JoinPointInvariants experiemnts
I defined the invariants both as a `Fixpoint` and as an `Inductive`, and
then proved them equal. This unrooted some bugs.

It also showed that efficient proving will depend on having good
induction principles: Right now the proof from isJoinPointsValid
(Fixpoint) to JoinPointInvariants (Inductive) itself is a bare `fix`,
and it was very fiddly to make sure I use the induction hypotheses only
ever on the right arguments, down to making sure that when I have `e0 =
e` in the contex, I apply them to the right one of these...

This does not seem to be tenable, so I will try to define some induction
schemes over CoreExpr next.