Outline how a proof about `go` would look like
if one is careful about not doing zeta-reduction (local definitions),
more of the Haskell code structure remains and one can actually prove
case by step. The `j_10__` definitions are actually useful here!

This proof will need a better story about termination. We probably need
a height function for `CoreExpr`.

This is actually a proof about a monadic function with a state monad,
where I defined some combinators to proof that a monadic function is
preserving some invariant on the state. Not to bad, at least when the
invariant is not changing ... if it were more complicated we’d probably
have to reinvent hoare logic here.