Turn the new Admitted in Exitify back into Qeds
some rewrite lemmas, like `exprFreeVars_Var`, which got unwanted
preconditions, do not have the precondition when we state them as
impliciations in the one direction that we need them.

But the whole business around `isLocalVar` needs more thought. It may be
that the `WellScoped` predicate is not yet what we want.