Verify Set.fromList
triggered by https://github.com/nomeata/containers-verified/issues/3 I
had a look at `fromList` again and understood that is is essentially the
same as `fromDistinctAscList`, just that is checks the invariant and
falling back to a naive `fromList'`. So based the former proof I was
able to conclude the latter.

There is a lot of redundancy between `fromDistinctAscList`,
`fromDistinctDescList` and` `fromList`, but I do not see how to easily
remove that.