No NCore!
I noticed that the `NCore` data type was essentially only used to encode
the preconditions of `Exitify`’s `go`, to be used in its termination proof
and other “low-level” proofs. I thus ditched `NCore` and instead
introduced `GoDom`, an inductive predicate that describes those
expressions that `go` is defined on. The resulting proof is a bit
shorter (130 lines), and hopefully easier to understand, with less types
around.