Make definitions depend on notations
Now, the Coq definition

    Definition f := ... op_zpzp__ ... .

in module `M` will depend on all of `M.op_zpzp__`, `++`, and `_++_`.  The first
is old news; the latter two are new.  They aren't qualified because notations
aren't qualified.

This exposes `Notation`s to `order` edits; note that the names they bind are
unqualified.

This change also makes notations binders, as when the pretty-printer was changed
to output the infix versions of the names it also changed the property that
making them non-binders relied on.