Solve FMapInterface type mismatch issue
This commit should address the

    Error: Signature components for label equal_1 do not match: expected type
    "??" but found type "??".

error.  The problem was that some of the general ordering/list theorems were
implicitly using in-context variables they shouldn't have been.  My best guess
is through `assumption`, but I'm not sure.  Transitively, `equal_1` and
`equal_2` used these, and so had extra arguments, causing the error message.  I
solved this by adding `clear` to the starts of those functions that were
accidentally picking up extra variables, but it might be a sign that being more
parsimonious with the section-level variables would be a good idea.

The fact that the error message was that egregiously terrible, however, is a Coq
bug and should be reported.