Drop support for `inline mutual f : ty -> pe`
We still smash types into the binders of fixpoints, but we set types with `set
type` now.  This doesn't provide a substitute for the "strict smushing"
behavior, where `inline mutual f : ty -> pe` would fail hard if the `ty`
couldn't be moved down to the binders, but I think that's okay.